Added new section <purchase>

+++ PointOfSaleAPI/Models/OrderItemModel.cs
@@ -6,8 +6,8 @@
         public int OrderId { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
-        public Decimal UnitPrice { get; set; }
-        public Decimal TotalPrice { get; set; }
+        public int UnitPrice { get; set; }
+        public int TotalPrice { get; set; }
         public virtual OrderModel OrderModels { get; set; }
         public virtual ProductModel ProductModels { get; set; }
     }
